chore: simplify codebase - remove dead code, dedup services/tests, fix layer violations
Removes dead code (unused NavigationMapper.IsEncerradas, InvestmentScopeParser.TryNormalize,
RepositoryConfigurationKeys.LocalJsonDefaultFileName), a hardcoded machine path, and duplicated
test doubles. Collapses the over-engineered Active/Historic BrokerBreakdown and
PortfolioAssetSummary service pairs into single scope-parametrized services matching the
existing SummaryService pattern, merging their duplicated test files. Moves the asset-class
validation rule out of AssetPricesController (was duplicated from the fetchers) and moves P&L
calculations out of the WPF Presentation layer into Financial.Domain/Application (mirroring the
XIRR extraction), adding the unit test coverage that was previously missing. Centralizes
duplicated percent-formatting logic in the web frontend.
