+++ Financial.Application/DependencyInjection/ApplicationServiceCollectionExtensions.cs
@@ -16,14 +16,11 @@ public static class ApplicationServiceCollectionExtensions
         services.AddSingleton<ITransactionService>(sp => sp.GetRequiredService<TransactionService>());
         services.AddSingleton<ITransactionQueryService>(sp => sp.GetRequiredService<TransactionService>());
         services.AddSingleton<IDividendService, DividendService>();
-        services.AddSingleton<IActiveBrokerBreakdownService, ActiveBrokerBreakdownService>();
-        services.AddSingleton<IHistoricBrokerBreakdownService, HistoricBrokerBreakdownService>();
         services.AddSingleton<IBrokerBreakdownService, BrokerBreakdownService>();
-        services.AddSingleton<IActivePortfolioAssetSummaryService, ActivePortfolioAssetSummaryService>();
-        services.AddSingleton<IHistoricPortfolioAssetSummaryService, HistoricPortfolioAssetSummaryService>();
         services.AddSingleton<IPortfolioAssetSummaryService, PortfolioAssetSummaryService>();
         services.AddSingleton<ISummaryService, SummaryService>();
         services.AddSingleton<IXirrCalculationService, XirrCalculationService>();
+        services.AddSingleton<IProfitCalculationService, ProfitCalculationService>();
 
         return services;
     }
