--- Financial.Application/Interfaces/IActiveBrokerBreakdownService.cs
@@ -1,8 +1,0 @@
-using Financial.Application.DTOs;
-
-namespace Financial.Application.Interfaces;
-
-public interface IActiveBrokerBreakdownService
-{
-    IReadOnlyList<PortfolioBreakdownItemDTO> GetBrokerBreakdown(string brokerName);
-}
