--- Financial.Application/Interfaces/IHistoricBrokerBreakdownService.cs
@@ -1,8 +1,0 @@
-using Financial.Application.DTOs;
-
-namespace Financial.Application.Interfaces;
-
-public interface IHistoricBrokerBreakdownService
-{
-    IReadOnlyList<PortfolioBreakdownItemDTO> GetBrokerBreakdown(string brokerName);
-}
