--- Financial.Application/Interfaces/IHistoricPortfolioAssetSummaryService.cs
@@ -1,8 +1,0 @@
-using Financial.Application.DTOs;
-
-namespace Financial.Application.Interfaces;
-
-public interface IHistoricPortfolioAssetSummaryService
-{
-    IReadOnlyList<PortfolioAssetSummaryItemDTO> GetPortfolioAssetsSummary(string brokerName, string portfolioName);
-}
