+++ Financial.Application/Interfaces/IProfitCalculationService.cs
@@ -1,0 +1,8 @@
+namespace Financial.Application.Interfaces;
+
+public interface IProfitCalculationService
+{
+    bool HasCostBasis(decimal averagePrice, decimal quantity);
+    decimal CalculateResultFraction(decimal averagePrice, decimal quantity, decimal currentValue);
+    decimal? CalculateProfitPercent(decimal currentValue, decimal costBasis);
+}
