+++ Financial.Application/Validation/InvestmentScopeParser.cs
@@ -4,9 +4,6 @@ namespace Financial.Application.Validation;
 
 public static class InvestmentScopeParser
 {
-    public static bool TryNormalize(string? value, out string normalized) =>
-        EnumParser.TryNormalize<InvestmentScope>(value, out normalized);
-
     public static bool TryParse(string? value, out InvestmentScope scope) =>
         EnumParser.TryParseEnum(value, out scope);
