+++ Financial.Infrastructure/Configuration/RepositoryConfigurationKeys.cs
@@ -4,7 +4,6 @@ public static class RepositoryConfigurationKeys
 {
     public const string Provider = "Repository:Provider";
     public const string LocalJsonDataFile = "DataJsonFile";
-    public const string LocalJsonDefaultFileName = "data.json";
     public const string GoogleDriveCredentialsPath = "GoogleDrive:CredentialsPath";
     public const string GoogleDriveFilePath = "GoogleDrive:FilePath";
 }
