+++ Tests/Financial.Infrastructure.Tests/Services/CryptocurrencyAssetPriceFetcherTests.cs
@@ -105,35 +105,4 @@ public class CryptocurrencyAssetPriceFetcherTests
         act.Should().Throw<InvalidOperationException>().WithMessage("*Coinbase*");
     }
 
-    private sealed class StubRepository : IRepository
-    {
-        private readonly List<Broker> _brokers;
-
-        public StubRepository(IEnumerable<Broker> brokers)
-        {
-            _brokers = brokers.ToList();
-        }
-
-        public IEnumerable<Asset> GetAssetsByBroker(string name, InvestmentScope scope = InvestmentScope.Active) => throw new NotImplementedException();
-
-        public IEnumerable<Asset> GetAssetsByBrokerPortfolio(string broker, string portfolio, InvestmentScope scope = InvestmentScope.Active) => throw new NotImplementedException();
-
-        public IEnumerable<Broker> GetBrokerList(InvestmentScope scope = InvestmentScope.Active) => _brokers;
-
-        public Asset? GetAsset(string brokerName, string portfolioName, string assetName, InvestmentScope scope = InvestmentScope.Active) => throw new NotImplementedException();
-
-        public Task SaveChangesAsync() => throw new NotImplementedException();
-    }
-
-    private sealed class FakeFinanceService : IFinanceService
-    {
-        private readonly AssetValueSnapshot? _snapshot;
-
-        public FakeFinanceService(AssetValueSnapshot? snapshot = null)
-        {
-            _snapshot = snapshot;
-        }
-
-        public AssetValueSnapshot GetAssetValue(AssetValueRequest request) => _snapshot ?? throw new NotImplementedException();
-    }
 }
